requirements.txt
eel
plexapi
configparser
pyatv==0.15.1
Flask
requests
gunicorn
pywebostv
flask-socketio
eventlet
pytz
python-dotenv
pexpect
PyQt6
PyQt6-WebEngine
appdirs
websocket-client
wakeonlan
setuptools==71.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.2.0
autocommand==2.2.2
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
chacha20poly1305-reuseable==0.13.2
charset-normalizer==3.4.0
click==8.1.7
configparser==7.1.0
cryptography==44.0.0
dnspython==2.7.0
Eel==0.18.1
eventlet==0.38.0
filetype==1.2.0
Flask==3.1.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mediafile==0.13.0
miniaudio==1.61
more-itertools==10.5.0
multidict==6.1.0
mutagen==1.47.0
packaging==24.2
pexpect==4.9.0
PlexAPI==4.16.0
propcache==0.2.1
protobuf==5.29.1
ptyprocess==0.7.0
pyatv==0.15.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.3
PyQt6-WebEngineSubwheel-Qt6==6.7.3
PyQt6_sip==13.9.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
pywebostv==0.8.9
requests==2.32.3
setuptools==71.0.0
simple-websocket==1.1.0
six==1.17.0
srptools==1.0.1
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
ws4py==0.5.1
wsproto==1.2.0
yarl==1.18.3
zeroconf==0.136.2
zope.event==5.0
zope.interface==7.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==24.2.0
autocommand==2.2.2
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
chacha20poly1305-reuseable==0.13.2
charset-normalizer==3.4.0
click==8.1.7
configparser==7.1.0
cryptography==44.0.0
dnspython==2.7.0
Eel==0.18.1
eventlet==0.38.0
filetype==1.2.0
Flask==3.1.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mediafile==0.13.0
miniaudio==1.61
more-itertools==10.5.0
multidict==6.1.0
mutagen==1.47.0
packaging==24.2
pexpect==4.9.0
PlexAPI==4.16.0
propcache==0.2.1
protobuf==5.29.1
ptyprocess==0.7.0
pyatv==0.15.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-WebEngine==6.7.0
PyQt6-WebEngine-Qt6==6.7.3
PyQt6-WebEngineSubwheel-Qt6==6.7.3
PyQt6_sip==13.9.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
pywebostv==0.8.9
requests==2.32.3
setuptools==71.0.0
simple-websocket==1.1.0
six==1.17.0
srptools==1.0.1
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
ws4py==0.5.1
wsproto==1.2.0
yarl==1.18.3
zeroconf==0.136.2
zope.event==5.0
zope.interface==7.2
